+++ src/platform_impl/windows/event_loop.rs
@@ -853,9 +853,16 @@ unsafe extern "system" fn public_window_callback<T: 'static>(
                     winuser::MonitorFromRect(&new_rect, winuser::MONITOR_DEFAULTTONULL);
                 match fullscreen {
                     Fullscreen::Borderless(ref mut fullscreen_monitor) => {
-                        if new_monitor != ptr::null_mut()
-                            && new_monitor != fullscreen_monitor.as_ref().unwrap().inner.hmonitor()
-                        {
+                        if new_monitor == ptr::null_mut() {
+                            return 0;
+                        }
+
+                        let current_monitor = fullscreen_monitor
+                            .as_ref()
+                            .map(inner.hmonitor())
+                            .unwrap_or(new_monitor);
+
+                        if new_monitor != current_monitor {
                             if let Ok(new_monitor_info) = monitor::get_monitor_info(new_monitor) {
                                 let new_monitor_rect = new_monitor_info.rcMonitor;
                                 window_pos.x = new_monitor_rect.left;