+++ src/window.rs
@@ -624,14 +624,12 @@ impl Window {
 
     /// Gets the window's current fullscreen state.
     ///
-    /// If `None` in `Fullscreen::Borderless` was returned it means
-    /// that the window is in a `Fullscreen` state, however there's
-    /// no monitor information available.
-    ///
     /// ## Platform-specific
     ///
     /// - **iOS:** Can only be called on the main thread.
     /// - **Android:** Will always return `None`.
+    /// - **Wayland** Can return `Borderless(None)` in some scenarios
+    ///
     #[inline]
     pub fn fullscreen(&self) -> Option<Fullscreen> {
         self.window.fullscreen()
@@ -857,8 +855,7 @@ impl Default for CursorIcon {
 /// Fullscreen mode to be set on certain output.
 ///
 /// Providing `None` to `Borderless` will result in fullscreen
-/// being set on a system preferred monitor. In general it'll be
-/// the current monitor.
+/// being set on a current monitor.
 #[derive(Clone, Debug, PartialEq)]
 pub enum Fullscreen {
     Exclusive(VideoMode),